% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxtix_tranz.R
\name{maxtix_tranz}
\alias{maxtix_tranz}
\alias{eigen_module}
\alias{positive_matrix_calc}
\alias{transitive_closure}
\alias{matrix_transitive_join}
\alias{consonanse_dissonanse}
\alias{cross_consonanse}
\alias{cross_dissonanse}
\alias{cross_positive_influence}
\alias{cross_negative_influence}
\alias{impuls_vector}
\alias{multiply_vector}
\alias{multiply_matrix}
\alias{maximum_matrix}
\alias{ik_pos_maximum}
\alias{ik_neg_maximum}
\alias{reverse_task}
\alias{direct_task}
\title{matrix_tranz}
\usage{
tnorm_functions

snorm_functions

snorm_functions_reverse

tnorm_functions_reverse

positive_matrix_calc(initmatrix)

transitive_closure(
  positivematrix,
  tnorm,
  snorm,
  snormMatrix,
  gammaTnormMean,
  algaTnorm,
  gammaTnorm,
  piTnorm,
  gammaSnorm,
  piSnorm
)

matrix_transitive_join(matrix, snorm, gammaSnorm, piSnorm)

consonanse_dissonanse(finalmatrix)

cross_consonanse(finalmatrix)

cross_dissonanse(finalmatrix)

cross_positive_influence(finalmatrix)

cross_negative_influence(finalmatrix)

impuls_vector(vector, matrix)

multiply_vector(matrix, vector)

multiply_matrix(
  matrix_1,
  matrix_2,
  tnorm,
  snorm,
  gammaTnormMean,
  algaTnorm,
  gammaTnorm,
  piTnorm,
  gammaSnorm,
  piSnorm
)

maximum_matrix(matrix_1, matrix_2)

ik_pos_maximum(matrix, initMatrix, ipath, jpath)

ik_neg_maximum(matrix, initMatrix, ipath, jpath)

reverse_task(
  df_matrix,
  vectorY,
  tnorm,
  tnorm_reverse,
  snorm,
  snormMatrix,
  snorm_reverse
)

direct_task(df_matrix, vectorX, tnorm, snorm, snormMatrix)
}
\arguments{
\item{initmatrix, df_matrix, snormMatrix, matrix, initMatrix, matrix_1, matrix_2, positivematrix, finalmatrix}{matrix}

\item{tnorm, gammaTnormMean, algaTnorm, gammaTnorm, piTnorm, gammaSnorm, piSnorm, snorm, tnorm_reverse, snorm_reverse}{norm}

\item{ipath, jpath}{parameter}

\item{vectorY, vector, vectorX}{vector}

\item{tnorm_functions, snorm_functions, snorm_functions_reverse, tnorm_functions_reverse}{function}
}
\description{
The maxtix_tranz set of functions is aimed to calculate dissonance, consonance and influence
}
