% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxtix_tranz_probability.R
\name{maxtix_tranz_probability}
\alias{maxtix_tranz_probability}
\alias{multiply_matrix_prob}
\alias{transitive_closure_prob}
\alias{transitive_closure_prob_max}
\alias{probability_matrix_transitive}
\title{FCM package with functions for matrix probability calculations}
\usage{
multiply_matrix_prob(matrix_1, matrix_2)

transitive_closure_prob(positivematrix)

transitive_closure_prob_max(positivematrix)

probability_matrix_transitive(maxmatrix)
}
\arguments{
\item{matrix_1}{matrix}

\item{matrix_2}{matrix}

\item{positivematrix}{matrix}

\item{maxmatrix}{matrix}
}
\value{
multiplication of matrix \code{matrix_1} and \code{matrix_2}

transitive closure of matrix \code{positivematrix}

transitive closure of matrix \code{positivematrix} with max function

joined transitive closure of matrix \code{maxmatrix}
}
\description{
The maxtix_tranz_probability set of functions is aimed to calculate maximum matrix based on transitive closure
}
