% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polinomial_eq_dominant.R
\name{polinomial_eq_dominant}
\alias{polinomial_eq_dominant}
\alias{calc_reverse_task}
\title{FCM package with functions for matrix manipulations}
\usage{
calc_reverse_task(matrix, vector, tnorm, tnorm_reverse, snorm, snorm_reverse)
}
\arguments{
\item{matrix}{matrix}

\item{vector}{vector}

\item{tnorm, tnorm_reverse, snorm, snorm_reverse}{norm}
}
\description{
polinomial_eq_dominant set contains 1 function: reverse task calculation.
Each function takes a matrix, vector and t-norm as arguments and returns a solution matrix.
The calculation procedure of the function includes a solution existence check and a solution check.
}
