% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_norm.R
\docType{data}
\name{s_norm}
\alias{s_norm}
\alias{drastic_sum_snorm}
\alias{bounded_sum_snorm}
\alias{einstein_sum_snorm}
\alias{algebraic_sum_snorm}
\alias{hamacher_sum_snorm}
\alias{max_snorm}
\alias{hamacher_union_operator_snorm}
\alias{yager_union_operator_snorm}
\alias{snorm_functions}
\alias{get_snorm}
\title{s_norm}
\format{
An object of class \code{list} of length 8.
}
\usage{
drastic_sum_snorm(element1, element2, gammaSnorm, piSnorm)

bounded_sum_snorm(element1, element2, gammaSnorm, piSnorm)

einstein_sum_snorm(element1, element2, gammaSnorm, piSnorm)

algebraic_sum_snorm(element1, element2, gammaSnorm, piSnorm)

hamacher_sum_snorm(element1, element2, gammaSnorm, piSnorm)

max_snorm(element1, element2, gammaSnorm, piSnorm)

hamacher_union_operator_snorm(element1, element2, gammaSnorm, piSnorm)

yager_union_operator_snorm(element1, element2, gammaSnorm, piSnorm)

snorm_functions

get_snorm(typeSnorm)
}
\arguments{
\item{element1, element2}{paramater}

\item{gammaSnorm, piSnorm, typeSnorm}{norm}
}
\description{
s_norm set of functions is aimed to calculate
drastic, einstein, algebraic, hamacher products, min and bounded difference S-norms
}
\keyword{datasets}
