% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_norm_reverse.R
\docType{data}
\name{s_norm_reverse}
\alias{s_norm_reverse}
\alias{drastic_sum_snorm_reverse}
\alias{bounded_sum_snorm_reverse}
\alias{einstein_sum_snorm_reverse}
\alias{algebraic_sum_snorm_reverse}
\alias{hamacher_sum_snorm_reverse}
\alias{max_snorm_reverse}
\alias{hamacher_union_operator_snorm_reverse}
\alias{yager_union_operator_snorm_reverse}
\alias{snorm_functions_reverse}
\alias{get_snorm_reverse}
\title{FCM package with functions for reverse S-norms calculations}
\format{
An object of class \code{list} of length 8.
}
\usage{
drastic_sum_snorm_reverse(element1, element2, gammaSnorm, piSnorm)

bounded_sum_snorm_reverse(element1, element2, gammaSnorm, piSnorm)

einstein_sum_snorm_reverse(element1, element2, gammaSnorm, piSnorm)

algebraic_sum_snorm_reverse(element1, element2, gammaSnorm, piSnorm)

hamacher_sum_snorm_reverse(element1, element2, gammaSnorm, piSnorm)

max_snorm_reverse(element1, element2, gammaSnorm, piSnorm)

hamacher_union_operator_snorm_reverse(element1, element2, gammaSnorm, piSnorm)

yager_union_operator_snorm_reverse(element1, element2, gammaSnorm, piSnorm)

snorm_functions_reverse

get_snorm_reverse(typeSnorm)
}
\arguments{
\item{element1, element2}{paramater}

\item{gammaSnorm, piSnorm, typeSnorm}{norm}
}
\description{
s_norm_reverse set of functions is aimed to calculate
drastic, einstein, algebraic, hamacher products, min and bounded difference reverse S-norms
}
\keyword{datasets}
