% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_norm.R
\docType{data}
\name{t_norm}
\alias{t_norm}
\alias{min_tnorm}
\alias{hamacher_product_tnorm}
\alias{algebraic_product_tnorm}
\alias{einstein_product_tnorm}
\alias{bounded_difference_tnorm}
\alias{drastic_product_tnorm}
\alias{parameterized_mean_intersection_operator_tnorm}
\alias{dubois_intersection_operator_tnorm}
\alias{hamacher_intersection_operator_tnorm}
\alias{yager_intersection_operator_tnorm}
\alias{tnorm_functions}
\alias{get_tnorm}
\title{t_norm}
\format{
An object of class \code{list} of length 10.
}
\usage{
min_tnorm(element1, element2, gammaTnormMean, algaTnorm, gammaTnorm, piTnorm)

hamacher_product_tnorm(
  element1,
  element2,
  gammaTnormMean,
  algaTnorm,
  gammaTnorm,
  piTnorm
)

algebraic_product_tnorm(
  element1,
  element2,
  gammaTnormMean,
  algaTnorm,
  gammaTnorm,
  piTnorm
)

einstein_product_tnorm(
  element1,
  element2,
  gammaTnormMean,
  algaTnorm,
  gammaTnorm,
  piTnorm
)

bounded_difference_tnorm(
  element1,
  element2,
  gammaTnormMean,
  algaTnorm,
  gammaTnorm,
  piTnorm
)

drastic_product_tnorm(
  element1,
  element2,
  gammaTnormMean,
  algaTnorm,
  gammaTnorm,
  piTnorm
)

parameterized_mean_intersection_operator_tnorm(
  element1,
  element2,
  gammaTnormMean,
  algaTnorm,
  gammaTnorm,
  piTnorm
)

dubois_intersection_operator_tnorm(
  element1,
  element2,
  gammaTnormMean,
  algaTnorm,
  gammaTnorm,
  piTnorm
)

hamacher_intersection_operator_tnorm(
  element1,
  element2,
  gammaTnormMean,
  algaTnorm,
  gammaTnorm,
  piTnorm
)

yager_intersection_operator_tnorm(
  element1,
  element2,
  gammaTnormMean,
  algaTnorm,
  gammaTnorm,
  piTnorm
)

tnorm_functions

get_tnorm(typeTnorm)
}
\arguments{
\item{element1, element2}{paramater}

\item{gammaTnormMean, algaTnorm, gammaTnorm, piTnorm, typeTnorm}{norm}
}
\description{
t_norm set of functions is aimed to calculate
drastic, einstein, algebraic, hamacher products, min and bounded difference T-norms
}
\keyword{datasets}
