% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_norm_reverse.R
\docType{data}
\name{t_norm_reverse}
\alias{t_norm_reverse}
\alias{min_tnorm_reverse}
\alias{hamacher_product_tnorm_reverse}
\alias{algebraic_product_tnorm_reverse}
\alias{einstein_product_tnorm_reverse}
\alias{bounded_difference_tnorm_reverse}
\alias{drastic_product_tnorm_reverse}
\alias{parameterized_mean_intersection_operator_tnorm_reverse}
\alias{dubois_intersection_operator_tnorm_reverse}
\alias{hamacher_intersection_operator_tnorm_reverse}
\alias{yager_intersection_operator_tnorm_reverse}
\alias{tnorm_functions_reverse}
\alias{get_tnorm_reverse}
\title{t_norm_reverse}
\format{
An object of class \code{list} of length 10.
}
\usage{
min_tnorm_reverse(
  element1,
  element2,
  gammaTnormMean,
  algaTnorm,
  gammaTnorm,
  piTnorm
)

hamacher_product_tnorm_reverse(
  element1,
  element2,
  gammaTnormMean,
  algaTnorm,
  gammaTnorm,
  piTnorm
)

algebraic_product_tnorm_reverse(
  element1,
  element2,
  gammaTnormMean,
  algaTnorm,
  gammaTnorm,
  piTnorm
)

einstein_product_tnorm_reverse(
  element1,
  element2,
  gammaTnormMean,
  algaTnorm,
  gammaTnorm,
  piTnorm
)

bounded_difference_tnorm_reverse(
  element1,
  element2,
  gammaTnormMean,
  algaTnorm,
  gammaTnorm,
  piTnorm
)

drastic_product_tnorm_reverse(
  element1,
  element2,
  gammaTnormMean,
  algaTnorm,
  gammaTnorm,
  piTnorm
)

parameterized_mean_intersection_operator_tnorm_reverse(
  element1,
  element2,
  gammaTnormMean,
  algaTnorm,
  gammaTnorm,
  piTnorm
)

dubois_intersection_operator_tnorm_reverse(
  element1,
  element2,
  gammaTnormMean,
  algaTnorm,
  gammaTnorm,
  piTnorm
)

hamacher_intersection_operator_tnorm_reverse(
  element1,
  element2,
  gammaTnormMean,
  algaTnorm,
  gammaTnorm,
  piTnorm
)

yager_intersection_operator_tnorm_reverse(
  element1,
  element2,
  gammaTnormMean,
  algaTnorm,
  gammaTnorm,
  piTnorm
)

tnorm_functions_reverse

get_tnorm_reverse(typeTnorm)
}
\arguments{
\item{element1, element2}{paramater}

\item{gammaTnormMean, algaTnorm, gammaTnorm, piTnorm, typeTnorm}{norm}
}
\description{
t_norm_reverse set of functions is aimed to calculate
drastic, einstein, algebraic, hamacher products, min and bounded difference reverse T-norms
}
\keyword{datasets}
